myapps/tribe/requirements.txt
## Graphs and Network Analysis
networkx==1.11
python-louvain==0.3

## Visualization
matplotlib==1.5.1
#pygraphviz==1.3.1

## Utilities
confire==0.2.0
python-dateutil==2.5.0
unicodecsv==0.14.1

## Other Dependencies
# networkX Dependencies
#scipy==0.17.0
decorator==4.0.9
# confire Dependencies
PyYAML==3.11
six==1.10.0
# matplotlib Dependencies
numpy==1.10.4
pytz==2015.7
cycler==0.10.0
pyparsing==2.1.0

## Testing (uncomment for development)
#nose==1.3.7
#mock==1.3.0
#coverage==4.0.3
#funcsigs==0.4
#pbr==1.8.1

## Packaging Requirements
#Python==2.7.10
#pip==8.1.0
#wheel==0.29.0
#setuptools==20.2.2
#wsgiref==0.1.2
